+++ Changelog.md
@@ -7,8 +7,9 @@
 - New: More blocks now support the Disguise Module: Keypad Door, Keypad Trapdoor, Passcode-Protected Chest, Reinforced Dispenser, Reinforced Dropper, Reinforced Hopper, Scanner Door, Scanner Trapdoor, Security Camera, Sonic Security System
 - Change: Security Camera head display
 - API: New interface IDisguisable to abstract away from the DisguisableBlock class. This means blocks no longer need to extend it to be disguisable, but instead need to implement the interfaces
-- Fix: Disguised secure redstone interfaces have incorrect collision, and are see-through
 - Fix: Crash when reading the SecurityCraft Manual
+- Fix: Disguised secure redstone interfaces have incorrect collision, and are see-through
+- Fix: The Secure Redstone Interface screen cannot be closed by pressing the inventory key
 
 --------------------------Changelog for v1.9.10 of SecurityCraft--------------------------